ItemsTabs completed.

+++ ObjectOrientedPractics/ObjectOrientedPractics/Services/ValueValidator.cs
@@ -31,7 +31,7 @@ namespace ObjectOrientedPractics.Services
         /// <param name="min">Левый предел промежутка.</param>
         /// <param name="max">Правый предел промежутка.</param>
         /// <param name="propertyName">Название свойства, в котором был вызван метод.</param>
-        public static void AssertValueInRange(double value, int min, int max, string propertyName)
+        public static void AssertValueInRange(double value, double min, double max, string propertyName)
         {
             if ((value < min) || (value > max))
             {
